agentic_ai_rag_tester/requirements.txt
openai>=1.61.0
pydantic>=2.0.0
PyYAML>=6.0.0
playwright>=1.49.0
beautifulsoup4>=4.12.0
requests>=2.30.0
pandas>=2.0.0
loguru>=0.7.0
streamlit>=1.40.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
fastapi>=0.111.0
uvicorn>=0.29.0
openpyxl>=3.1.2
xlrd>=2.0.1
pypdf>=4.1.0
python-docx>=1.1.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
langchain-text-splitters>=0.2.0
langchain>=0.2.0
